+++ controller/src/main_valveCtrl.h
@@ -37,7 +37,7 @@ namespace control {
         //Diese Funktion wird vom StringBuilder aufgerufen und sammelt die Daten der Ventile
         //um sie in den Ausgabestring zu schreiben
         //Die Funktion fragt die einzelnen Ventile nach ihren Werten ab
-        void getValveValueList(uint16_t valveValueList[]);
+        void getValveValueList(uint8_t valveValueList[]);
         uint16_t getAmount_valve();
         //wird von Main_LabCom kontinuierlich abgefragt, ob die queue abgeschlossen ist.
         bool getQueueFinished();
